+++ test/harness.go
@@ -4,16 +4,29 @@ package test
 import (
 	"fmt"
 	"log"
+	"net"
+	"net/http"
 	"os"
 	"path/filepath"
+	"regexp"
+	"runtime"
 	"strconv"
+	"strings"
 	"time"
 
+	"github.com/disney/quanta/custom/functions"
+	proxy "github.com/disney/quanta/quanta-proxy-lib"
 	"github.com/disney/quanta/server"
 	"github.com/disney/quanta/shared"
+	"github.com/disney/quanta/sink"
+	"github.com/disney/quanta/source"
 	"github.com/hashicorp/consul/api"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 
 	u "github.com/araddon/gou"
+	"github.com/araddon/qlbridge/expr"
+	"github.com/araddon/qlbridge/expr/builtins"
+	"github.com/araddon/qlbridge/schema"
 )
 
 // Setup - Initialize test harness
@@ -74,7 +87,7 @@ func StartNodes(nodeStart int, nodeCount int) (*server.Node, error) {
 		bindAddr := "127.0.0.1"
 		port := 4000 + index
 
-		consul := "127.0.0.1:8500"
+		consul := bindAddr + ":8500"
 
 		memLimit := 0
 
@@ -171,4 +184,109 @@ func StartNodes(nodeStart int, nodeCount int) (*server.Node, error) {
 	// for {
 	// 	time.Sleep(1 * time.Second)
 	// }
+}
+
+func StartProxy(count int) {
+
+	// for index := 0; index < count; index++ { // TODO: more than one proxy
+	index := 0
+
+	proxy.SetupCounters()
+	proxy.Init()
+
+	logging := "DEBUG"
+	environment := "DEV"
+	Version := "1.0.0"
+	proxy.ConsulAddr = "127.0.0.1:8500"
+	// cognito url for token service publicKeyURL := "" // unused
+	proxy.QuantaPort = 4000
+	proxyHostPort := 4040 + index
+
+	// region := "us-east-1"
+
+	if strings.ToUpper(logging) == "DEBUG" || strings.ToUpper(logging) == "TRACE" {
+		if strings.ToUpper(logging) == "TRACE" {
+			expr.Trace = true
+		}
+		u.SetupLogging("debug")
+	} else {
+		shared.InitLogging(logging, environment, "Proxy", Version, "Quanta")
+	}
+
+	go func() {
+		// Initialize Prometheus metrics endpoint.
+		http.Handle("/metrics", promhttp.Handler())
+		http.ListenAndServe(":2112", nil)
+	}()
+
+	log.Printf("Connecting to Consul at: [%s] ...\n", proxy.ConsulAddr)
+	consulConfig := &api.Config{Address: proxy.ConsulAddr}
+	errx := shared.RegisterSchemaChangeListener(consulConfig, proxy.SchemaChangeListener)
+	if errx != nil {
+		u.Error(errx)
+		os.Exit(1)
+	}
+
+	poolSize := 3
+
+	// If the pool size is not configured then set it to the number of available CPUs
+	// this is weird atw
+	sessionPoolSize := poolSize
+	if sessionPoolSize == 0 {
+		sessionPoolSize = runtime.NumCPU()
+		log.Printf("Session Pool Size not set, defaulting to number of available CPUs = %d", sessionPoolSize)
+	} else {
+		log.Printf("Session Pool Size = %d", sessionPoolSize)
+	}
+
+	// Match 2 or more whitespace chars inside string
+	reWhitespace := regexp.MustCompile(`[\s\p{Zs}]{2,}`)
+	_ = reWhitespace
+
+	// load all of our built-in functions
+	builtins.LoadAllBuiltins()
+	sink.LoadAll()      // Register output sinks
+	functions.LoadAll() // Custom functions
+
+	// sess, errx := session.NewSession(&aws.Config{
+	// 	Region: aws.String(region),
+	// })
+	// if errx != nil {
+	// 	u.Error(errx)
+	// 	os.Exit(1)
+	// }
+	// _ = sess
+	// metrics = cloudwatch.New(sess)
+
+	// configDir := "../test/testdata/config"
+	//configDir := "../shared/testdata/config"
+	configDir := "../configuration" // do we really want this here?
+
+	// Construct Quanta source
+	// it's just one for a whole pool.
+	//for i := 0; i < sessionPoolSize; i++ {
+	src, err := source.NewQuantaSource(configDir, proxy.ConsulAddr, proxy.QuantaPort+1, sessionPoolSize) // do we really want this here?
+	if err != nil {
+		u.Error(err)
+	}
+	//schema.RegisterSourceAsSchema("quanta-node-"+strconv.Itoa(i), src)
+	schema.RegisterSourceAsSchema("quanta", src)
+
+	// Start server endpoint
+	portStr := strconv.FormatInt(int64(proxyHostPort), 10)
+	l, err := net.Listen("tcp", "127.0.0.1:"+portStr)
+	if err != nil {
+		panic(err.Error())
+	}
+
+	go func() {
+		for {
+			conn, err := l.Accept()
+			if err != nil {
+				u.Errorf(err.Error())
+				return
+			}
+			go proxy.OnConn(conn)
+		}
+	}()
 }
